Implemented ship movement in Stylized Water 2 asset
This ship doesn't use the buoyancy and floating transform script already given in the Stylized Water 2 asset as it heavily interfered the position and rotation of the ship.

Used just buoyancy script for the ship. Other stationary objects have the floating transform script that float naturally in the ocean surface.

Remaining to fix the joystick on left hand bottom side so it can be made responsive for all platforms.